Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0224-basic-calculator/0224-basic-calculator.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int calculate(string s) {
         stack<int> st;
-        int result = 0;
+        long long int result = 0;
         long long int number = 0;
         int sign = 1;
